+++ lib/features/shift/domain/usecases/open_shift_usecase.dart
@@ -8,7 +8,7 @@ class OpenShiftUseCase {
   final ShiftRepository repository;
   OpenShiftUseCase(this.repository);
 
-  Future<Either<Failure, Shift>> execute({required double startingCash, required int cashierId}) async {
+  Future<Either<Failure, Shift>> execute({required int startingCash, required int cashierId}) async {
     if (startingCash < 0) return const Left(CacheFailure('العهدة لا يمكن أن تكون بالسالب'));
     return await repository.openShift(startingCash: startingCash, cashierId: cashierId);
   }
